+++ Common/src/main/java/at/petrak/hexcasting/api/casting/eval/vm/FrameForEach.kt
@@ -1,6 +1,5 @@
 package at.petrak.hexcasting.api.casting.eval.vm
 
-import at.petrak.hexcasting.api.casting.SpellList
 import at.petrak.hexcasting.api.casting.eval.CastResult
 import at.petrak.hexcasting.api.casting.eval.ResolvedPatternType
 import at.petrak.hexcasting.api.casting.iota.Iota
@@ -11,7 +10,6 @@ import at.petrak.hexcasting.common.lib.hex.HexEvalSounds
 import com.mojang.serialization.MapCodec
 import com.mojang.serialization.codecs.RecordCodecBuilder
 import net.minecraft.network.RegistryFriendlyByteBuf
-import net.minecraft.network.codec.ByteBufCodecs
 import net.minecraft.network.codec.StreamCodec
 import net.minecraft.server.level.ServerLevel
 
@@ -25,18 +23,16 @@ import net.minecraft.server.level.ServerLevel
  * @property acc concatenated list of final stack states after each iteration
  */
 data class FrameForEach(
-    val data: SpellList,
-    val code: SpellList,
-    val contextStack: List<Iota>,
-    val stashedStack: List<Iota>,
+    val data: TreeList<Iota>,
+    val code: TreeList<Iota>,
+    val contextStack: TreeList<Iota>,
+    val stashedStack: TreeList<Iota>,
     val acc: TreeList<Iota>
 ) : ContinuationFrame {
 
     /** When halting, we add the stack state at halt to the stack accumulator, then return the stashed stack, plus the accumulator. */
-    override fun breakDownwards(stack: List<Iota>): Pair<Boolean, List<Iota>> {
-        val newStack = stashedStack.toMutableList()
-        newStack.add(ListIota(acc.appendedAll(stack)))
-        return true to newStack
+    override fun breakDownwards(stack: TreeList<Iota>): Pair<Boolean, TreeList<Iota>> {
+        return true to stashedStack.appended(ListIota(acc.appendedAll(stack)))
     }
 
     /** Step the Thoth computation, enqueueing one code evaluation. */
@@ -49,22 +45,18 @@ data class FrameForEach(
         val newAcc = acc.appendedAll(harness.image.stack)
 
         // If we still have data to process...
-        val (newStack, newImage, newCont) = if (data.nonEmpty) {
-            // Restore the context stack,
-            val stack = contextStack.toMutableList()
-            // push the next datum to the top of the stack,
-            stack.add(data.car)
+        val (newStack, newImage, newCont) = if (!data.isEmpty()) {
+            // Create a stack composed of the context stack plus the next datum
+            val stack = contextStack.appended(data.head())
             val cont2 = continuation
                 // put the next Thoth object back on the stack for the next Thoth cycle,
-                .pushFrame(FrameForEach(data.cdr, code, contextStack, stashedStack, newAcc))
+                .pushFrame(FrameForEach(data.tail(), code, contextStack, stashedStack, newAcc))
                 // and prep the Thoth'd code block for evaluation.
                 .pushFrame(FrameEvaluate(code, true))
             Triple(stack, harness.image.withUsedOp(), cont2)
         } else {
-            // Else, restore the stashed stack,
-            val stack = stashedStack.toMutableList()
-            // and dump our final list onto the stack.
-            stack.add(ListIota(newAcc))
+            // Else, restore the stashed stack, and dump our final list on top.
+            val stack = stashedStack.appended(ListIota(newAcc))
             Triple(stack, harness.image, continuation)
         }
         return CastResult(
@@ -78,7 +70,7 @@ data class FrameForEach(
         )
     }
 
-    override fun size() = data.size() + code.size() + acc.size + contextStack.size + stashedStack.size
+    override fun size() = data.size + code.size + acc.size + contextStack.size + stashedStack.size
 
     override val type: ContinuationFrame.Type<*> = TYPE
 
@@ -87,26 +79,23 @@ data class FrameForEach(
         val TYPE: ContinuationFrame.Type<FrameForEach> = object : ContinuationFrame.Type<FrameForEach> {
             val CODEC = RecordCodecBuilder.mapCodec<FrameForEach> { inst ->
                 inst.group(
-                    SpellList.CODEC.fieldOf("data").forGetter { it.data },
-                    SpellList.CODEC.fieldOf("code").forGetter { it.code },
-                    IotaType.TYPED_CODEC.listOf().fieldOf("context").forGetter { it.contextStack },
-                    IotaType.TYPED_CODEC.listOf().fieldOf("stashed").forGetter { it.stashedStack },
-                    IotaType.TYPED_CODEC.listOf().fieldOf("accumulator").forGetter { it.acc }
+                    TreeList.codecOf(IotaType.TYPED_CODEC).fieldOf("data").forGetter { it.data },
+                    TreeList.codecOf(IotaType.TYPED_CODEC).fieldOf("code").forGetter { it.code },
+                    TreeList.codecOf(IotaType.TYPED_CODEC).fieldOf("context").forGetter { it.contextStack },
+                    TreeList.codecOf(IotaType.TYPED_CODEC).fieldOf("stashed").forGetter { it.stashedStack },
+                    TreeList.codecOf(IotaType.TYPED_CODEC).fieldOf("accumulator").forGetter { it.acc }
                 ).apply(inst) { a, b, c, d, e ->
-                    FrameForEach(a, b, c, d, TreeList.from(e))
+                    FrameForEach(a, b, c, d, e)
                 }
             }
             val STREAM_CODEC = StreamCodec.composite(
-                SpellList.STREAM_CODEC, FrameForEach::data,
-                SpellList.STREAM_CODEC, FrameForEach::code,
-                IotaType.TYPED_STREAM_CODEC
-                    .apply(ByteBufCodecs.list()), FrameForEach::contextStack,
-                IotaType.TYPED_STREAM_CODEC
-                    .apply(ByteBufCodecs.list()), FrameForEach::stashedStack,
-                IotaType.TYPED_STREAM_CODEC
-                    .apply(ByteBufCodecs.list()), FrameForEach::acc
+                IotaType.TYPED_STREAM_CODEC.apply(TreeList.streamCodecOp()), FrameForEach::data,
+                IotaType.TYPED_STREAM_CODEC.apply(TreeList.streamCodecOp()), FrameForEach::code,
+                IotaType.TYPED_STREAM_CODEC.apply(TreeList.streamCodecOp()), FrameForEach::contextStack,
+                IotaType.TYPED_STREAM_CODEC.apply(TreeList.streamCodecOp()), FrameForEach::stashedStack,
+                IotaType.TYPED_STREAM_CODEC.apply(TreeList.streamCodecOp()), FrameForEach::acc
             ) { a, b, c, d, e ->
-                FrameForEach(a, b, c, d, TreeList.from(e))
+                FrameForEach(a, b, c, d, e)
             }